# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BARINCAIRO.COM API",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/v1/venues": {
      "get": {
        "tags": [
          "Venues"
        ],
        "summary": "List Venues Geojson",
        "operationId": "list_venues_geojson_api_v1_venues_get",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "min_lng,min_lat,max_lng,max_lat",
              "title": "Bbox"
            },
            "description": "min_lng,min_lat,max_lng,max_lat"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category slug filter",
              "title": "Category"
            },
            "description": "Category slug filter"
          },
          {
            "name": "price_range",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Price range filter e.g. $, $$, $$$",
              "title": "Price Range"
            },
            "description": "Price range filter e.g. $, $$, $$$"
          },
          {
            "name": "vibe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Vibe tag slug filter",
              "title": "Vibe"
            },
            "description": "Vibe tag slug filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/venues/{slug}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "summary": "Get Venue Detail",
        "operationId": "get_venue_detail_api_v1_venues__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeature"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Feature",
            "title": "Type",
            "description": "GeoJSON feature type",
            "default": "Feature"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoJSONGeometry"
          },
          "properties": {
            "$ref": "#/components/schemas/VenueProperties"
          }
        },
        "type": "object",
        "required": [
          "geometry",
          "properties"
        ],
        "title": "GeoJSONFeature"
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "description": "GeoJSON collection type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features",
            "description": "List of GeoJSON features"
          }
        },
        "type": "object",
        "title": "GeoJSONFeatureCollection"
      },
      "GeoJSONGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Point",
            "title": "Type",
            "description": "GeoJSON geometry type",
            "default": "Point"
          },
          "coordinates": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Coordinates",
            "description": "[longitude, latitude]"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "GeoJSONGeometry"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VenueProperties": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Venue primary key ID"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug",
            "description": "Unique venue slug"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name",
            "description": "Venue name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description"
          },
          "address": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Address",
            "description": "Physical address"
          },
          "working_hours": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Hours",
            "description": "Working hours description"
          },
          "price_range": {
            "type": "string",
            "enum": [
              "$",
              "$$",
              "$$$",
              "$$$$"
            ],
            "title": "Price Range",
            "description": "Price level indicator",
            "default": "$$"
          },
          "vibe_description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Vibe Description",
            "description": "Short vibe summary"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url",
            "description": "Primary photo URL"
          },
          "google_maps_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Maps Url",
            "description": "Google Maps URL for directions"
          },
          "category_slug": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Category Slug",
            "description": "Category slug"
          },
          "category_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Category Name",
            "description": "Category display name"
          },
          "vibes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Vibes",
            "description": "List of associated vibe tag slugs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "address",
          "category_slug",
          "category_name"
        ],
        "title": "VenueProperties"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "barincairo_auth"
      }
    }
  }
}